\name{eduplicates}
\alias{eduplicates}
\title{Identifies dublicates essential genes}
\usage{
eduplicates(essential)
}
\arguments{
  \item{essential}{The dataframe with information on
  essential genes.}
}
\value{
A overview of duplicated essential genes and their parrent
scaffold.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

