\name{ephylum}
\alias{ephylum}
\title{Extracts a subset of scaffolds}
\usage{
ephylum(eg)
}
\arguments{
  \item{eg}{Pre-formatted file of genomes and}
}
\value{
A list with two objects. $plot a ggplot object, $table a
table with summery data.
}
\description{
A nice long description
}
\examples{
phylum_data <- ephylum(eg)
phylum_data$plot
phylum_data$table
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

