\name{extract}
\alias{extract}
\title{Extracts a subset of scaffolds}
\usage{
extract(data, seletion)
}
\arguments{
  \item{data}{The dataframe containing all data.}

  \item{selection}{The subspace to extract.}
}
\value{
The subset of scaffolds in the original dataframe within
the defined selection.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

