\name{extract_nodes}
\alias{extract_nodes}
\title{Extract all scaffolds connected by paired reads}
\usage{
extract_nodes(original, subset)
}
\arguments{
  \item{original}{The original dataframe with all scaffolds
  (d).}

  \item{subset}{The subset of extracted scaffolds.}
}
\value{
A list with nodes and scaffolds
}
\description{
A nice long description.
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

