\name{ggplot_locator}
\alias{ggplot_locator}
\title{Interactively extract points ggplot2 figures}
\usage{
ggplot_locator(p)
}
\arguments{
  \item{p}{(required). A ggplot2 object.}
}
\value{
X and Y coordinates marked in the plot.
}
\description{
Adds "locator" functionallity to plots made with ggplot2
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

