\name{import_network}
\alias{import_network}
\title{Imports a connection network and adds attributes}
\usage{
import_network(network, data, nconnections, nsize, ewidth)
}
\arguments{
  \item{network}{Flat network file.}

  \item{data}{Data frame of scaffold information.}

  \item{nconnections}{Minumum number of connections.}

  \item{nsize}{Scale for node size.}

  \item{ewidth}{Scale for edge width.}
}
\value{
A igraph object with scaffold attributes attached.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

