\name{import_rrna}
\alias{import_rrna}
\title{Imports and formats silva rRNA results}
\usage{
import_rrna(data, type)
}
\arguments{
  \item{data}{The dataframe containing all data.}

  \item{type}{Either 16S or 23S. Just adds the header.}
}
\value{
Nicely formatted dataframe.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

