\name{merge_network}
\alias{merge_network}
\title{Identifies clusters of scaffolds from a graph}
\usage{
merge_network(graph, data)
}
\arguments{
  \item{graph}{A igraph object.}

  \item{data}{Data frame of scaffold information.}
}
\value{
A data frame with scaffold cluster information attached.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

