% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmadd.R
\name{mmadd}
\alias{mmadd}
\title{Add data to any mmgenome object}
\usage{
mmadd(data, new)
}
\arguments{
\item{data}{(required) The existing mmgenome object.}

\item{newdata}{(required) The new dataframe that is to be added.}
}
\value{
A list with 2 dataframes: scaffolds and essential (if provided).
}
\description{
Add new dataframes to an excisting mmgenome object.
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
