\name{mmexport}
\alias{mmexport}
\title{Exports scaffolds in fasta format}
\usage{
mmexport(data, assembly, file)
}
\arguments{
  \item{data}{The dataframe containing all data on
  scaffolds.}

  \item{assembly}{The assembly.}

  \item{file}{Name of output file.}
}
\value{
An overview of key statistics of the scaffolds.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

