% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmformat_esom.R
\name{mmformat_esom}
\alias{mmformat_esom}
\title{Format raw esom data for use in mmgenome}
\usage{
mmformat_esom(names, bm, umx)
}
\arguments{
\item{names}{(required) The .names esom file.}

\item{bm}{(required) The .bm esom file.}

\item{umx}{(required) The .umx esom file.}

\item{xname}{Name of the esom x-coordinate (default: "esomx")}

\item{yname}{Name of the esom y-coordinate (default: "esomy")}
}
\value{
A list with two dataframes. The esom scaffold coordinates and the esom map boundaries.
}
\description{
Format raw esom data generated using databionics ESOM software.
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
