% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmformat_ppsp.R
\name{mmformat_ppsp}
\alias{mmformat_ppsp}
\title{Format ppsp data for use in mmgenome}
\usage{
mmformat_ppsp(data)
}
\arguments{
\item{data}{(required) The PP.pOUT file from PPSP.}

\item{ranks}{A vector of which phylogenetic ranks to keep (default: c("phylum", "class))}

\item{ranknames}{A vector with names of the kept ranks.}
}
\value{
A dataframe with cleaned PPSP data.
}
\description{
Format PhyloPythiaS+ data for use in mmgenome
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
