\docType{package}
\name{mmgenome}
\alias{mmgenome}
\alias{mmgenome-package}
\title{mmgenome: multi-metagenome}
\description{
Tools for extracting individual genomes from metagenomes
}
\section{Individual functions}{
  \itemize{ \item \code{\link{mmload}}: Merge data
  function.  \item \code{\link{mmplot}}: Default plot
  function.  \item \code{\link{mmplot_selction}}:
  Highlights a selection made with ggplot_locater on a
  mmplot.  \item \code{\link{mmplot_locator}}: Adds
  "locator" functionallity to plots made with ggplot2.
  \item \code{\link{mmplot_pairs}}: Pairs plot.  \item
  \code{\link{mmplot_network}}: Network plot.  \item
  \code{\link{mmextract}}: Extracts a subset of scaffolds
  based on input from \code{\link{ggplot_locator}}.  \item
  \code{\link{mmextract_network}}: Extracts PE connected
  scaffolds.  \item \code{\link{mmstats}}: An overview of
  key statistics of the scaffolds.  \item
  \code{\link{mmstats_duplicates}}: Identifies duplicated
  essential single copy genes.  \item
  \code{\link{mmref_especies}}: Reference data on essential
  genes.  \item \code{\link{mmref_ephylum}}: Reference data
  on essential genes on phylum level. }
}

