% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmimport.R
\name{mmimport}
\alias{mmimport}
\title{Import data}
\usage{
mmimport(file)
}
\arguments{
\item{file}{The Rmarkdown file that contains the load definitions.}

\item{save}{Save an image of the importated data for fast loading (default: T).}

\item{force}{Forces re-import of data if an previous image is present (default: F).}

\item{save.filename}{Filename of the image to save or load (default: "data.RData")}
}
\value{
Loaded data as detailed in the Rmarkdown file.
}
\description{
Imports data through a Rmarkdown template.
}
\examples{

\dontrun{
mmimport(file = "Load_data.Rmd")
}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
