% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mmplot}
\alias{mmplot}
\title{Function to visualise metagenomes}
\usage{
mmplot(data, x, y)
}
\arguments{
\item{data}{(required) A dataframe with scaffold information.}

\item{x}{(required) x-axis variable.}

\item{y}{(required) y-axis variable.}

\item{color}{Color by a specific variable (default: phylum)}

\item{log.x}{log10 scale the x-axis (default: T)}

\item{log.y}{log10 scale the y-axis (default: T)}

\item{log.color}{log10 scale the colors (default: F)}

\item{minlength}{Minimum length of plotted scaffolds.}

\item{network}{Network used to plot connections between scaffolds.}

\item{nconnections}{Minimum number of connections to plot (default: 0).}

\item{duplicates}{Mark scaffolds with duplicated essential genes (default: F)}

\item{labels}{If scaffold names are to be plotted (default: F)}

\item{resize}{Rescale the size of the scaffolds (default: 1)}
}
\value{
a ggplot2 object
}
\description{
Versitile plotting function to visualise metagnome assemblies loaded through \code{mmload}. Wraps ggplot2 for pretty default plots.
}
\examples{
\dontrun{
data(rocco)
mmplot(data = d, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "phylum", minlength = 10000)
}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

