\name{mmplot_locator}
\alias{mmplot_locator}
\title{Interactively extract points from ggplot2 objects}
\usage{
mmplot_locator(p, round)
}
\arguments{
  \item{p}{(required) A ggplot2 object.}

  \item{nround}{number of digits}
}
\value{
X and Y coordinates marked in the plot.
}
\description{
Adds "locator" functionallity to plots made with ggplot2
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

