\name{mmplot_selection}
\alias{mmplot_selection}
\title{Wraps ggplot2 for pretty default plots with subset highlighted}
\usage{
mmplot_selection(data, x, y, logx, logy, selection)
}
\arguments{
  \item{plot}{A plot generated by mmplot.}

  \item{sel}{A subset selection of the mmplot.}
}
\value{
a ggplot2 object
}
\description{
A nice long description.
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

