% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmref.R
\name{mmref}
\alias{mmref}
\title{Compare essential genes in a bin to a group of references}
\usage{
mmref(data, tax.level, tax.compare)
}
\arguments{
\item{data}{The dataframe containing the genome bin of interest (optional).}

\item{tax.level}{What level to compare on. Supports Kingdom, Phylum, Class, Genus, Species (default: Phylum)}

\item{tax.compare}{A vector with names on what should be used for comparison (e.g. c("Proteobacteria", "Actinobacteria")) or "all".}

\item{tax.aggregate}{Summarise the data to a fixed taxonomic rank. Supports Kingdom, Phylum, Class, Genus, Species (default: tax.level).}

\item{summarise}{Report "mean" or "median" HMM counts (default: mean).}

\item{output}{Either complete or plot (default: plot)}

\item{plot.display}{Show all HMMs even if they are not contained in the selected references (default: T).}

\item{plot.sort}{Sort the plot (default: T)}
}
\value{
An overview of essential genes in the bin and selected reference genomes.
}
\description{
Compare essential genes in an extracted genome bin to a group of references
}
\examples{

\dontrun{
mmref(tax.level = "Phylum", tax.compare = c("Proteobacteria", "Actinobacteria")) 

res <- mmref(tax.level = "Phylum", tax.compare = c("Proteobacteria", "Actinobacteria"), output = "complete") 
res$table

mmref(tax.level = "Phylum", tax.compare = "all") 
mmref(tax.level = "Phylum", tax.compare = "Proteobacteria", tax.aggregate = "Class")
mmref(tax.level = "Phylum", tax.compare = "Proteobacteria", tax.aggregate = "Class", summarise = "median") 
mmref(tax.level = "Class", tax.compare = "Betaproteobacteria", tax.aggregate = "Strain", summarise = "median")

mmref(tax.level = "Phylum", tax.compare = "Proteobacteria", tax.aggregate = "Class", summarise = "median", data = dA)}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
