% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstats.R
\name{mmstats}
\alias{mmstats}
\title{Calculates summary statistics on a set of scaffolds}
\usage{
mmstats(data, ncov)
}
\arguments{
\item{data}{The dataframe containing the bin data (e.g. dA).}

\item{ncov}{The number of coverage datasets (default = 2).}

\item{data_compare}{The dataframe containing the all data (e.g. d).}
}
\value{
An overview of key statistics of the scaffolds.
}
\description{
Calculates summary statistics on a set of scaffolds. Can be used on any object loaded using \code{mmload}.
}
\examples{

\dontrun{
data(rocco)
mmstats(data = d, ncov = 4)
}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
