\name{mmstats_duplicates}
\alias{mmstats_duplicates}
\title{Identifies duplicates essential genes}
\usage{
mmstats_duplicates(data)
}
\arguments{
  \item{data}{The dataframe.}
}
\value{
A overview of duplicated essential genes and their parrent
scaffold.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

