% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsubset.R
\name{mmsubset}
\alias{mmsubset}
\title{Subset scaffolds}
\usage{
mmsubset(data, ...)
}
\arguments{
\item{data}{(required) The dataframe containing all data.}

\item{...}{(required) Further arguments are passed on to the subset function.}
}
\value{
The subset of scaffolds.
}
\description{
Subset scaffolds using the default subset syntax.
}
\examples{

\dontrun{
data(rocco)

subset <- mmsubset(data = d, gc > 60)
}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
