\name{sstats}
\alias{sstats}
\title{Calculates statistics on a set of scaffolds}
\usage{
sstats(scaffolds, essential, ncov)
}
\arguments{
  \item{scaffolds}{The dataframe containing all data on
  scaffolds.}

  \item{essential}{The dataframe containing all data on
  essential genes.}

  \item{ncov}{The number of coverage datasets (default =
  2).}
}
\value{
An overview of key statistics of the scaffolds.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

