% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmexport.R
\name{mmexport}
\alias{mmexport}
\title{Exports scaffolds in fasta format}
\usage{
mmexport(data, assembly, file)
}
\arguments{
\item{data}{The dataframe containing all data.}

\item{assembly}{The assembly}

\item{file}{Name of output file.}
}
\value{
An overview of key statistics of the scaffolds.
}
\description{
Exports scaffolds in fasta format using the Biostings package.
}
\examples{

\dontrun{
data(rocco)

p <- mmplot(data = d, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "phylum", minlength = 3000)

sel <- data.frame(C13.12.03  =  c(1.39, 2.07, 16.8, 19.4, 7.72, 1.76),
                  C14.01.09  =  c(29.4, 67.6, 85.9, 43.6, 16.7, 14.9))
mmplot_selection(p, sel)

dA <- mmextract(d, sel)

mmexport(data=dA, assembly=assembly, file = "my_subset.fa")
}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
