% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmextract_network.R
\name{mmextract_network}
\alias{mmextract_network}
\title{Extract scaffolds connected by paired-end reads}
\usage{
mmextract_network(subset, network, original, nconnections, type)
}
\arguments{
\item{subset}{The subset dataframe to use for network extraction or a vector of scaffold names.}

\item{network}{The network of connected scaffold.}

\item{original}{The original dataframe (default: d).}

\item{nconnections}{The minumum number of connections needed (default: 2).}

\item{type}{Extract "direct" or "complete" (all) linked scaffolds (default: direct).}
}
\value{
A dataframe with scaffold information.
}
\description{
Extracts scaffolds that are connected by paired-end or mate-pair reads.
}
\examples{

\dontrun{
data(rocco)

p <- mmplot(data = d, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "phylum", minlength = 3000)
sel <- data.frame(C13.12.03  =  c(4.94, 8.99, 10.8, 7.22, 5.79),
                  C14.01.09  =  c(43.9, 57.4, 35.7, 25.7, 32.6))
mmplot_selection(p, sel) 

dA <- mmextract(d, sel)
mmplot(data = dA, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "phylum", minlength = 3000)
mmplot_network(data = dA, network = pe, color = "phylum", nconnections = 10)

dB <- mmextract_network(subset = dA, original = d, network = pe, nconnections = 10, type = "direct")
mmplot(data = dB, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "phylum", minlength = 3000)
mmplot_network(data = dB, network = pe, color = "phylum", nconnections = 10) 

}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
