\docType{package}
\name{mmgenome}
\alias{mmgenome}
\alias{mmgenome-package}
\title{mmgenome: multi-metagenome}
\description{
Tools for extracting individual genomes from metagenomes
}
\section{Individual functions}{
  \itemize{ \item \code{\link{ggplot_locator}}: Adds
  "locator" functionallity to plots made with ggplot2.
  \item \code{\link{extract}}: Extracts a subset of
  scaffolds based on input from
  \code{\link{ggplot_locator}}.  \item
  \code{\link{sstats}}: An overview of key statistics of
  the scaffolds.  \item \code{\link{clean_tax}}: Cleans
  consensus taxonomy files.  \item
  \code{\link{import_rrna}}: Import silva taxonomy
  information.  \item \code{\link{import_network}}: Import
  PE connection network.  \item
  \code{\link{merge_network}}: Merge connection network
  with scaffold information.  \item
  \code{\link{eduplicates}}: Identifies duplicated
  essential single copy genes.  \item
  \code{\link{extract_nodes}}: Extracts all PE connected
  scaffolds.  \item \code{\link{extract_linked}}: Extracts
  direct connected scaffolds.  \item
  \code{\link{especies}}: Reference data on essential
  genes.  \item \code{\link{ephylum}}: Reference data on
  essential genes on phylum level. }
}

