% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmgenome.R
\docType{package}
\name{mmgenome}
\alias{mmgenome}
\alias{mmgenome-package}
\title{mmgenome: multi-metagenome}
\description{
Tools for extracting individual genomes from metagenomes
}
\section{Individual functions}{


\itemize{
  \item \code{\link{mmload}}: Merge data function.
  \item \code{\link{mmimport}}: Import data function.
  \item \code{\link{mmplot}}: Default plot function.
  \item \code{\link{mmplot_selction}}: Highlights a selection made with mmplot_locater on a mmplot.
  \item \code{\link{mmplot_locator}}: Adds "locator" functionallity to ggplot2 objects.
  \item \code{\link{mmplot_pairs}}: Pairs plot.
  \item \code{\link{mmplot_network}}: Network plot.
  \item \code{\link{mmextract}}: Extracts a subset of scaffolds based on input from \code{\link{mmplot_locator}}.
  \item \code{\link{mmextract_network}}: Extracts PE/MP connected scaffolds.
  \item \code{\link{mmsubset}}: Subset scaffolds based on the normal subset syntax.
  \item \code{\link{mmstats}}: An overview of key statistics of the scaffolds.
  \item \code{\link{mmref}}: Compare essential gene content to selected references.
  \item \code{\link{mmexport}}: Export selected scaffolds to fasta file.
}
}

