% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmload.R
\name{mmload}
\alias{mmload}
\title{Data loading function}
\usage{
mmload(assembly, coverage, essential)
}
\arguments{
\item{assembly}{(required) The assembly.}

\item{coverage}{(required) A vector with the names of the coverage datasets.}

\item{essential}{A dataframe with information on essential genes.}

\item{pca}{Add PCA analyis of tetranucleotide frequencies (default: T)}

\item{pca.ncomp}{Number of principal components to return (default: 3).}

\item{tax}{A dataframe with taxonomic classification.}

\item{tax.freq}{Remove phyla seen less than X times (default: 20).}

\item{tax.clean}{Clean up some of the phyla names (default: T)}

\item{tax.expand}{Use class level assignments for specific phyla.}

\item{other}{A vector with the names of additional datasets.}
}
\value{
A list with 2 dataframes: scaffolds and essential (if provided).
}
\description{
This function is used to load information on scaffolds into a single object that is used by all subsequent mmgenome functions.
}
\examples{

\dontrun{
d <- mmload(assembly = assembly, 
            coverage = c("A", "B", "C", "D"), 
            essential = ess,           
            tax = tax,
            tax.expand = "Proteobacteria",
            tax.freq = 50,
            other = "rRNA16S"
            )
}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
