% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mmplot_locator.R
\name{mmplot_locator}
\alias{mmplot_locator}
\title{Interactively extract points from ggplot2 objects}
\usage{
mmplot_locator(p, sig)
}
\arguments{
\item{p}{(required) A ggplot2 object.}

\item{sig}{The number of significant digits (default: 3).}
}
\value{
X and Y coordinates marked in the plot.
}
\description{
Adds "locator" functionallity to plots made with mmplot (ggplot2).
}
\examples{
\dontrun{
data(rocco)

p <- mmplot(data = d, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "phylum", minlength = 3000)
p
sel <- mmplot_locator(p)
sel <- data.frame(C13.12.03  =  c(1.39, 2.07, 16.8, 19.4, 7.72, 1.76),
                  C14.01.09  =  c(29.4, 67.6, 85.9, 43.6, 16.7, 14.9))
mmplot_selection(p, sel)
}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

