% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot_locator.R
\name{mmplot_locator}
\alias{mmplot_locator}
\title{Interactively define a selection subspace on mmplot or mmplot_network}
\usage{
mmplot_locator(p, sig)
}
\arguments{
\item{p}{(required) A mmplot or mmplot_network object.}

\item{sig}{The number of significant digits (default: 3).}

\item{trans}{Specify x and y-axis transformation used in mmplot. The format is trans=c(x-axis, y-axis) i.e. trans=c("", "log10").  The following transformations are supported: No transformation (default) = "". Log10 = "log10". Square root transformation = "sqrt".}
}
\value{
X and Y coordinates marked in the plot or scaffold names.
}
\description{
Interactively define a selection subspace on \code{mmplot} and \code{mmplot_network}. The selection can be used to extract subsets of mmgenome data objects using \code{mmextract}. The function works on generic ggplot2 objects,
}
\examples{

\dontrun{
data(rocco)

p <- mmplot(data = d, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "pps_phylum", minlength = 5000)
sel <- mmplot_locator(p, trans=c("log10", "log10"))
sel <- data.frame(C13.12.03  =  c(1.39, 2.07, 16.8, 19.4, 7.72, 1.76),
                  C14.01.09  =  c(29.4, 67.6, 85.9, 43.6, 16.7, 14.9))
mmplot_selection(p, sel)

dA <- mmextract(d, sel)

pn <- mmplot_network(data = dA, network = pe, color = "essential", nconnections = 10)
sel <- mmplot_locator(pn)
sel <- c("3951", "6944")
mmplot_network(data = dA, network = pe, color = "essential", nconnections = 10, highlight= sel)
}
}
\author{
Soren M. Karst \email{sorenkarst@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
