\name{mmplot_pairs}
\alias{mmplot_pairs}
\title{Generates a pairs plot}
\usage{
mmplot_pairs(data, variables, color, nsize)
}
\arguments{
  \item{data}{The dataframe to plot.}

  \item{variables}{The vaiables to plot as a list.}

  \item{color}{Coloring of the plot (default: gc)}

  \item{nsize}{Resize the points (default: 1)}
}
\value{
A pairs plot object.
}
\description{
A nice long description
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

