% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot_pairs.R
\name{mmplot_pairs}
\alias{mmplot_pairs}
\title{Generates a ggplot2 style pairs plot}
\usage{
mmplot_pairs(data, variables)
}
\arguments{
\item{data}{(required) The dataframe to plot.}

\item{variables}{(required) The vaiables to plot as a vector.}

\item{color}{Coloring of the plot (default: gc).}

\item{log.color}{Log scale the color (default: F).}

\item{log}{A vector of variables to log scale.}

\item{minlength}{Minimum length of the plotted scaffolds.}

\item{resize}{Rescale the size of the scaffolds (default: 1).}

\item{textsize}{Size of the legend text (deafult = 8).}

\item{point.size}{Use a fixed size for points instead of scaffold length.}

\item{highlight}{Mark selected scaffolds on the plot. Either as a vector of scaffold names or as a full subset of data.}

\item{hightlight.color}{Color of the highlighted scaffolds (default: "darkred").}
}
\value{
A pairs plot object.
}
\description{
Generates a ggplot2 style pairs plot.
}
\examples{
\dontrun{
data(rocco)

mmplot_pairs(data = d, variables = c("C13.12.03", "C14.01.09", "C13.11.25", "gc", "PC2"), color = "essential", log = c("C13.12.03", "C14.01.09", "C13.11.25"), minlength = 10000)

}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
