% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot_selection.R
\name{mmplot_selection}
\alias{mmplot_selection}
\title{Highlight subsets}
\usage{
mmplot_selection(data, sel)
}
\arguments{
\item{plot}{A plot generated by mmplot.}

\item{sel}{Coordinates of the subset selection of the mmplot.}
}
\value{
a ggplot2 object
}
\description{
Highlights subsets made with \code{mmplot_locator} on a \code{mmplot} object.
}
\examples{

\dontrun{
data(rocco)

p <- mmplot(data = d, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "essential", minlength = 3000)
p
sel <- mmplot_locator(p)
sel <- data.frame(C13.12.03  =  c(1.39, 2.07, 16.8, 19.4, 7.72, 1.76),
                  C14.01.09  =  c(29.4, 67.6, 85.9, 43.6, 16.7, 14.9))
mmplot_selection(p, sel)
}
}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
